Add api for role business

+++ domain/src/main/java/com/wansensoft/dto/PageSizeDTO.java
@@ -24,11 +24,11 @@ public class PageSizeDTO {
     /**
      * 查询列表总记录数
      */
-    int page = 0;
+    Long page = 0L;
 
     /**
      * 每页显示条数，默认10
      */
-    int pageSize = 10;
+    Long pageSize = 10L;
 
 }
